requirements.txt
streamlit
plotly
pandas
numpy
firebase-admin
st-keyup
python-dotenv
requests 